requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.6
Cython==0.29.24
Django==3.2.8
django-cors-headers==3.8.0
djangorestframework==3.12.4
idna==3.2
numpy==1.21.2
pandas==1.2.5
pycodestyle==2.7.0
python-dateutil==2.8.2
pytz==2021.3
razorpay==1.2.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.7
